Changed to use ghostscript to convert pdf to photos instead of Wand
