+++ requirements.txt
@@ -1,3 +1,4 @@
+ghostscript>=0.6
 img2pdf>=0.3.3
 logbook>=1.4.3
 python-dotenv>=0.10.3
@@ -6,6 +7,4 @@ Pillow>=6.0.0
 PyPDF2>=1.26.0
 slackclient>=2.0.1
 textblob>=0.15.3
-Wand>=0.5.4
--e git+https://github.com/zeshuaro/pdf-diff.git#egg=pdf-diff
 git+https://github.com/zeshuaro/noteshrink.git#egg=noteshrink
